requirements.txt
discord.py~=2.5.2
python-steam-api~=2.1.1